Adds basic heatmap of daily activity
This is just a basic example of what the daily heatmap could be, there are some issues with it but I wanted everyone's thoughts on it before I continued working.

Currently, the activity is counted as "lessons_completed" I'm in agreement with
others in the issue that this metric may be a bit too much for most users to get
 a daily streak, the lessons are pretty bulky especially the challenges.

Currently there are some issues with the heat map:
* It does not work on mobile (it shows but hovering to see the date won't work).
* I couldn't figure out how to write a test for `lesson_completed_array`, I'd lo
ve to get some help on this.
* I'm not sure that displaying tooltips as a tooltip DIV is the best way to go a
bout this, I'm not sure how GitHub does it.  Using a tooltip div is a bit hacky
because it requires the div to be a direct child of the body tag, this is becaus
e of the necessity for absolute positioning for the tool tip to be displayed.