Fixes NodeJS badge for logged in users
Users who are logged in and have started the NodeJS track were seeing the default Odin Project logo.
The reason for this was that the badge had not been added to the `borderless_badges` section of
the course_decorator.  There was also some odd spacing around the SVG itsef but that should now
also be fixed in the SVG itself.